Move ProjectedNormal to directional.py
