+++ numpyro/distributions/continuous.py
@@ -1320,118 +1320,3 @@ class TruncatedPolyaGamma(Distribution):
     @classmethod
     def tree_unflatten(cls, aux_data, params):
         return cls(batch_shape=aux_data)
-
-
-class ProjectedNormal(TorchDistribution):
-    """
-    Projected isotropic normal distribution of arbitrary dimension.
-
-    This distribution over directional data is qualitatively similar to the von
-    Mises and von Mises-Fisher distributions, but permits tractable variational
-    inference via reparametrized gradients.
-
-    To use this distribution with autoguides, use ``poutine.reparam`` with a
-    :class:`~pyro.infer.reparam.projected_normal.ProjectedNormalReparam`
-    reparametrizer in the model, e.g.::
-
-        @poutine.reparam(config={"direction": ProjectedNormalReparam()})
-        def model():
-            direction = pyro.sample("direction",
-                                    ProjectedNormal(torch.zeros(3)))
-            ...
-
-    .. note:: This implements :meth:`log_prob` only for dimensions {2,3}.
-
-    [1] D. Hernandez-Stumpfhauser, F.J. Breidt, M.J. van der Woerd (2017)
-        "The General Projected Normal Distribution of Arbitrary Dimension:
-        Modeling and Bayesian Inference"
-        https://projecteuclid.org/euclid.ba/1453211962
-    """
-    arg_constraints = {"concentration": constraints.real_vector}
-    support = constraints.sphere
-    has_rsample = True
-
-    def __init__(self, concentration, *, validate_args=None):
-        assert concentration.dim() >= 1
-        self.concentration = concentration
-        batch_shape = concentration.shape[:-1]
-        event_shape = concentration.shape[-1:]
-        super().__init__(batch_shape, event_shape, validate_args=validate_args)
-
-    @property
-    def mean(self):
-        """
-        Note this is the mean in the sense of a centroid in the submanifold
-        that minimizes expected squared geodesic distance.
-        """
-        return safe_normalize(self.concentration)
-
-    @property
-    def mode(self):
-        return safe_normalize(self.concentration)
-
-    def rsample(self, sample_shape=torch.Size()):
-        shape = self._extended_shape(sample_shape)
-        x = self.concentration.new_empty(shape).normal_()
-        x = x + self.concentration
-        x = safe_normalize(x)
-        return x
-
-    def log_prob(self, value):
-        if self._validate_args:
-            event_shape = value.shape[-1:]
-            if event_shape != self.event_shape:
-                raise ValueError(f"Expected event shape {self.event_shape}, "
-                                 f"but got {event_shape}")
-            self._validate_sample(value)
-        dim = int(self.concentration.shape[-1])
-        if dim == 2:
-            return _projected_normal_log_prob_2(self.concentration, value)
-        if dim == 3:
-            return _projected_normal_log_prob_3(self.concentration, value)
-        raise NotImplementedError(
-            f"ProjectedNormal.log_prob() is not implemented for dim = {dim}. "
-            "Consider using poutine.reparam with ProjectedNormalReparam."
-        )
-
-
-def _projected_normal_log_prob_2(concentration, value):
-    def _dot(x, y):
-        return (x[..., None, :] @ y[..., None])[..., 0, 0]
-
-    # We integrate along a ray, factorizing the integrand as a product of:
-    # a truncated normal distribution over coordinate t parallel to the ray, and
-    # a univariate normal distribution over coordinate r perpendicular to the ray.
-    t = _dot(concentration, value)
-    t2 = t.square()
-    r2 = _dot(concentration, concentration) - t2
-    perp_part = (-0.5) * r2 - 0.5 * math.log(2 * math.pi)
-
-    # This is the log of a definite integral, computed by mathematica:
-    # Integrate[x/(E^((x-t)^2/2) Sqrt[2 Pi]), {x, 0, Infinity}]
-    # = (t + Sqrt[2/Pi]/E^(t^2/2) + t Erf[t/Sqrt[2]])/2
-    para_part = (jnp.exp((-0.5) * t2) * ((2 / math.pi) ** 0.5)
-                 + t * (1 + jnp.erf(t * 0.5 ** 0.5))).mul(0.5).log()
-
-    return para_part + perp_part
-
-
-def _projected_normal_log_prob_3(concentration, value):
-    def _dot(x, y):
-        return (x[..., None, :] @ y[..., None])[..., 0, 0]
-
-    # We integrate along a ray, factorizing the integrand as a product of:
-    # a truncated normal distribution over coordinate t parallel to the ray, and
-    # a bivariate normal distribution over coordinate r perpendicular to the ray.
-    t = _dot(concentration, value)
-    t2 = t.square()
-    r2 = _dot(concentration, concentration) - t2
-    perp_part = (-0.5) * r2 - math.log(2 * math.pi)
-
-    # This is the log of a definite integral, computed by mathematica:
-    # Integrate[x^2/(E^((x-t)^2/2) Sqrt[2 Pi]), {x, 0, Infinity}]
-    # = t/(E^(t^2/2) Sqrt[2 Pi]) + ((1 + t^2) (1 + Erf[t/Sqrt[2]]))/2
-    para_part = jnp.log(t * jnp.exp((-0.5) * t2) / (2 * math.pi) ** 0.5
-                        + (1 + t2) * (1 + jnp.erf(t * 0.5 ** 0.5)) / 2)
-
-    return para_part + perp_part
